# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Laya API",
    "description": "Run Laya typed decisions (choice / score / noul) over text, JSON objects or conversation turns, using your own questions or a built-in preset. Requests are auto-routed to the best checkpoint, or pinned with `model`.\n\nAuthenticate with an API key (`X-API-Key` or `Authorization: Bearer`) or HTTP Basic, whichever is configured on the server.",
    "version": "0.5.0"
  },
  "paths": {
...
    "/models": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "List checkpoints",
        "description": "Available checkpoints and which are currently resident.",
        "operationId": "list_models_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "503": {
            "description": "Model not loaded yet"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/presets": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "List built-in question presets",
        "description": "Return the ready-to-use question sets (triage/email/guard/moderation/router).",
        "operationId": "list_presets_presets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "object",
                  "title": "Response List Presets Presets Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "503": {
            "description": "Model not loaded yet"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/detect": {
      "post": {
        "tags": [
          "meta"
        ],
        "summary": "Detect script and language",
        "description": "Report script, best-effort language and `is_english` for a state (what routing uses).",
        "operationId": "detect_detect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Detect Detect Post"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "503": {
            "description": "Model not loaded yet"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/email/state": {
      "post": {
        "tags": [
          "meta"
        ],
        "summary": "Build a clean email state",
        "description": "Clean an email body and structure it as a state for `/predict`.",
        "operationId": "email_state_endpoint_email_state_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Email State Endpoint Email State Post"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "503": {
            "description": "Model not loaded yet"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/predict": {
      "post": {
        "tags": [
          "predict"
        ],
        "summary": "Predict one state",
        "description": "Run typed questions over a single state and return calibrated answers.",
        "operationId": "predict_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "503": {
            "description": "Model not loaded yet"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
    "/predict/bulk": {
      "post": {
        "tags": [
          "predict"
        ],
        "summary": "Predict many states",
        "description": "Run typed questions over many states.\n\nAccepts either `states` with shared `questions`/`preset`, or `items` where each\nstate can override its questions and model. Laya's public API predicts one state\nat a time (its internals can batch, but the public `Agent.predict` cannot), so\nthis loops. Per-state errors are isolated and returned inline.",
        "operationId": "predict_bulk_predict_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkPredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkPredictResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "503": {
            "description": "Model not loaded yet"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          },
          {
            "BasicAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkPredictResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/BulkItemResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "count",
          "results"
        ],
        "title": "BulkPredictResponse"
      },
...
      "DetectRequest": {
        "properties": {
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "state"
        ],
        "title": "DetectRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelsResponse": {
        "properties": {
          "available": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available"
          },
          "loaded": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Loaded"
          }
        },
        "type": "object",
        "required": [
          "available",
          "loaded"
        ],
        "title": "ModelsResponse"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "A key from the server's API_KEYS."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key sent as a bearer token."
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "A user:password from the server's BASIC_AUTH."
      }
    }
  },
...
}